pyproject.toml
[project]
name = "rag-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.4",
    "google-genai>=1.50.1",
    "ipykernel>=7.1.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "typing-extensions>=4.8.0",
    "unstructured>=0.10.0",
]
